# docs/openapi.yaml
openapi: 3.0.3
info:
  title: petowo
  version: 0.1.0
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create User
      operationId: create_user_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSchemaCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
        '401':
          description: No unauthorized client access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
    get:
      tags:
        - users
      summary: Get User All
      operationId: get_user_all_users_get
      parameters:
        - name: skip
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
            title: Skip
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            exclusiveMinimum: false
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSchema'
                title: Response Get User All User Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
  /users/{user_id}:
    get:
      tags:
        - users
      summary: Get User By Id
      operationId: get_user_by_id_users__user_id__get
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
    put:
      tags:
        - users
      summary: Update User
      operationId: update_user_users__user_id__post
      parameters:
        - name: user_id
          in: path
          description: User Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSchemaUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
        '401':
          description: No unauthorized client access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Users can be updated only by themself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
...
  /auth/jwt/authorization:
    post:
      tags:
        - authorization
      summary: Login
      operationId: authorization_jwt_login_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSchema'
        required: false
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthSchemaDetails'
                  - {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
...
components:
  schemas:
...
    AuthSchema:
      oneOf:
        - type: string
        - $ref: '#/components/schemas/AuthSchemaSignIn'
    AuthSchemaDetails:
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
      type: object
      required:
        - access_token
        - refresh_token
      title: AuthDetails
    AuthSchemaSignIn:
      properties:
        email:
          type: string
        password:
          type: string
          title: Password
        fingerprint:
          type: string
      type: object
      required:
        - email
        - password
        - fingerprint
      title: AuthSchemaSignIn
    HTTPError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/HTTPError'
          type: array
          title: Detail
      type: object
      title: HTTPError
...
    UserSchema:
      properties:
        id:
          type: integer
          minimum: 0
          title: id
        email:
          type: string
        role:
          type: string
        name:
          type: string
      type: object
      required:
        - id
        - email
        - role
        - name
      title: UserSchema
    UserSchemaCreate:
      properties:
        email:
          type: string
        hashed_password:
         type: string
        role:
          type: string
        name:
          type: string
      type: object
      required:
        - email
        - hashed_password
        - role
        - name
      title: UserSchemaCreate
    UserSchemaUpdate:
      properties:
        id:
          type: integer
          minimum: 0
          title: id
        email:
          type: string
        hashed_password:
          type: string
        role:
          type: string
        name:
          type: string
      type: object
      required:
        - id
        - email
        - hashed_password
        - role
        - name
      title: UserSchemaUpdate
...